WLTS-31 refactor constants and tests (part 2)
   [WLTS-31](https://tools.hmcts.net/jira/browse/WLTS-31)

Refactor controller constants and tests.
* Part 2: update references and remove obsolete tests.
