+++ src/integrationTest/java/uk/gov/hmcts/reform/translate/controllers/RootControllerIT.java
@@ -1,20 +1,17 @@
 package uk.gov.hmcts.reform.translate.controllers;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
+import uk.gov.hmcts.reform.translate.BaseTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest
-@Disabled
-class GetWelcomeTest {
+class RootControllerIT extends BaseTest {
 
     @Autowired
     private transient MockMvc mockMvc;
